+++ lua/LUA/ak/data/TrainsAndTracksStatePublisher.lua
@@ -1,27 +1,27 @@
-if AkDebugLoad then print("[#Start] Loading ak.data.TrainsAndTracksJsonCollector ...") end
+if AkDebugLoad then print("[#Start] Loading ak.data.TrainsAndTracksStatePublisher ...") end
 local TrainDetection = require("ak.data.TrainDetection")
 local TrainRegistry = require("ak.train.TrainRegistry")
 local RollingStockRegistry = require("ak.train.RollingStockRegistry")
 
-TrainsAndTracksJsonCollector = {}
+TrainsAndTracksStatePublisher = {}
 
 local enabled = true
 local initialized = false
-TrainsAndTracksJsonCollector.name = "ak.data.TrainsAndTracksJsonCollector"
+TrainsAndTracksStatePublisher.name = "ak.data.TrainsAndTracksStatePublisher"
 
 local data = {}
 
-function TrainsAndTracksJsonCollector.initialize()
+function TrainsAndTracksStatePublisher.initialize()
     if not enabled or initialized then return end
     TrainDetection.initialize()
 
     initialized = true
 end
 
-function TrainsAndTracksJsonCollector.collectData()
+function TrainsAndTracksStatePublisher.collectData()
     if not enabled then return end
 
-    if not initialized then TrainsAndTracksJsonCollector.initialize() end
+    if not initialized then TrainsAndTracksStatePublisher.initialize() end
     TrainDetection.update()
 
     TrainRegistry.fireChangeTrainsEvent()
@@ -30,4 +30,4 @@ function TrainsAndTracksJsonCollector.collectData()
     return data
 end
 
-return TrainsAndTracksJsonCollector
+return TrainsAndTracksStatePublisher